requirements.txt
requests==2.31.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
plotly==5.18.0
torch==2.1.2
mlflow==2.8.1
backtrader==1.9.78.123
ta==0.10.2
python-dotenv==1.0.0
pytest==7.4.3